scripts/requirements.txt
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-google-genai
chromadb
tiktoken
jq
accelerate
torch
transformers
huggingface_hub
autoawq
peft
trl
bitsandbytes
datasets
